+++ src/main/java/com/code/pojo/ItemItemset.java
@@ -1,13 +1,13 @@
 package com.code.pojo;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.Version;
-import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableField;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -25,11 +25,11 @@ public class ItemItemset implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "item_id", type = IdType.AUTO)
-    private Integer itemId;
+    private Long itemId;
 
-    private Integer itemsetId;
+    private Long itemsetId;
 
-    private Integer sortId;
+    private Long sortId;
 
     @TableField("passingRate")
     private Float passingRate;
